Issue 96: Move test assertions in better namespace
