+++ behat.yaml
@@ -50,11 +50,11 @@ end-to-end:
                 - Khatovar\Tests\EndToEnd\Context\User\CommonContext: ~
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\AdminContext: ~
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\DeleteContext:
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertUsersAreAdministrableOnes'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertUsersAreAdministrableOnes'
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\EditContext:
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertUsersAreAdministrableOnes'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertUsersAreAdministrableOnes'
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\ListContext:
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertUsersAreAdministrableOnes'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertUsersAreAdministrableOnes'
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\RoleContext:
                       - '@Khatovar\Component\User\Domain\Repository\UserRepositoryInterface'
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\StatusContext:
@@ -62,11 +62,11 @@ end-to-end:
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\SuperAdminContext: ~
                 - Khatovar\Tests\EndToEnd\Context\User\Administrator\ViewContext: ~
                 - Khatovar\Tests\EndToEnd\Context\User\Anonymous\AuthenticationContext:
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertAuthenticatedAsUser'
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertUserIsAnonymous'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertAuthenticatedAsUser'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertUserIsAnonymous'
                 - Khatovar\Tests\EndToEnd\Context\User\Anonymous\RegisterNewUserContext:
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertAuthenticatedAsUser'
-                      - '@Khatovar\Tests\EndToEnd\Service\Assert\AssertUserIsAnonymous'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertAuthenticatedAsUser'
+                      - '@Khatovar\Tests\EndToEnd\Assertion\AssertUserIsAnonymous'
                       - '@Khatovar\Component\User\Domain\Repository\UserRepositoryInterface'
                 - Khatovar\Tests\EndToEnd\Context\User\Anonymous\ResetPasswordContext: ~
                 - Khatovar\Tests\EndToEnd\Context\User\RegularUser\ProfileContext: ~
