+++ tests/end-to-end/Context/User/Anonymous/AuthenticationContext.php
@@ -23,9 +23,9 @@ declare(strict_types=1);
 
 namespace Khatovar\Tests\EndToEnd\Context\User\Anonymous;
 
+use Khatovar\Tests\EndToEnd\Assertion\AssertAuthenticatedAsUser;
+use Khatovar\Tests\EndToEnd\Assertion\AssertUserIsAnonymous;
 use Khatovar\Tests\EndToEnd\Context\User\UserRawContext;
-use Khatovar\Tests\EndToEnd\Service\Assert\AssertAuthenticatedAsUser;
-use Khatovar\Tests\EndToEnd\Service\Assert\AssertUserIsAnonymous;
 
 /**
  * @author Damien Carcel <damien.carcel@gmail.com>
